pyproject.toml
[project]
name = "ib-flex-query"
version = "0.1.0"
description = "A clean, modern Python client for downloading Interactive Brokers flex statements with custom date range support"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "IB Flex Query Client" }
]
keywords = ["interactive-brokers", "flex-query", "trading", "finance", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
